LayoutTests/fast/forms/programmatic-focus-after-displaying-parent-expected.txt

This test verifies that programmatically focusing a textarea in a newly displayed container changes the active element. To test manually, tap or click below.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.activeElement is textarea
PASS didFinishActivatingElement became true
PASS successfullyParsed is true

TEST COMPLETE
